+++ BackEnd/app/Http/Controllers/JobControllers/RegJobsController.php
@@ -2,13 +2,16 @@
 
 namespace App\Http\Controllers\JobControllers;
 
+use App\Models\User;
 use App\Models\RegJob;
 use App\Models\Company;
 use App\Filters\JobFilter;
 use App\Models\Individual;
 use Illuminate\Http\Request;
+use App\Policies\RegJobPolicy;
 use App\Models\RegJobCompetetor;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Resources\RegJobResource;
 use App\Http\Requests\AddRegJobRequest;
 use App\Http\Resources\RegJobCollection;
@@ -20,18 +23,20 @@ class RegJobsController extends Controller
 {
     public function PostRegJob(AddRegJobRequest $request)
     {
+        
         $validated = $request->validated();
 
         // Get user
         $user = auth()->user();
-        $company = Company::where('user_id', $user->id)->get();
+        
+        $policy = new RegJobPolicy();
 
-        // Check user
-        if ($company == null) {
+        if (!$policy->PostRegJob(User::find($user->id))) {
             return response()->json([
-                'errors' => ['user' => 'Invalid user']
+                'errors' => ['user' => 'Unauthorized']
             ], 401);
         }
+        $company=Company::where('user_id',$user->id);
 
         $validated['company_id'] = $company->id;
         $job = RegJob::create($validated);
@@ -63,6 +68,15 @@ class RegJobsController extends Controller
     }
     public function DeleteRegJob(Request $request,RegJob $regJob)
     {
+        $user=Auth::user();
+        $policy = new RegJobPolicy();
+
+        if (!$policy->DeleteRegJob(User::find($user->id),$regJob)) {
+            return response()->json([
+                'errors' => ['user' => 'Unauthorized']
+            ], 401);
+        }
+
         $regJob->delete();
         
         return response()->json([
@@ -72,13 +86,15 @@ class RegJobsController extends Controller
     public function ApplyRegJob(ApplyRegJobRequest $request){
         // Get user
         $user = auth()->user();
-        $individual=Individual::where('user_id',$user->id);
-        // Check user
-        if ($individual == null) {
+        $policy = new RegJobPolicy();
+
+        if (!$policy->ApplyRegJob(User::find($user->id))) {
             return response()->json([
-                'errors' => 'Invalid user'
+                'errors' => ['user' => 'Unauthorized']
             ], 401);
         }
+
+        $individual=Individual::where('user_id',$user->id);
         $validated['individal_id']=$individual->id;
         $validated=$request->validated();
         $RegJobCompetetor= RegJobCompetetor::create($validated);
